- add prev_invalid_sql and prev_error_msg to prompt - shift generate_prompt to utils
