+++ eval/llama_cpp_runner.py
@@ -3,32 +3,13 @@ from concurrent.futures import ThreadPoolExecutor, as_completed
 
 from eval.eval import compare_query_results
 import pandas as pd
-from utils.pruning import prune_metadata_str
+from utils.gen_prompt import generate_prompt
 from utils.questions import prepare_questions_df
 from utils.creds import db_creds_all
 from tqdm import tqdm
 from time import time
 from utils.reporting import upload_results
 from llama_cpp import Llama
-
-
-def generate_prompt(
-    prompt_file, question, db_name, instructions="", k_shot_prompt="", public_data=True
-):
-    with open(prompt_file, "r") as f:
-        prompt = f.read()
-    question_instructions = question + " " + instructions
-
-    pruned_metadata_str = prune_metadata_str(
-        question_instructions, db_name, public_data
-    )
-    prompt = prompt.format(
-        user_question=question,
-        instructions=instructions,
-        table_metadata_string=pruned_metadata_str,
-        k_shot_prompt=k_shot_prompt,
-    )
-    return prompt
 
 
 def process_row(llm, row):
@@ -101,7 +82,7 @@ def run_llama_cpp_eval(args):
     for prompt_file, output_file in zip(prompt_file_list, output_file_list):
         # create a prompt for each question
         df["prompt"] = df[
-            ["question", "db_name", "instructions", "k_shot_prompt"]
+            ["question", "db_name", "instructions", "k_shot_prompt", "glossary", "table_metadata_string", "prev_invalid_sql", "prev_error_msg"]
         ].apply(
             lambda row: generate_prompt(
                 prompt_file,
@@ -109,6 +90,10 @@ def run_llama_cpp_eval(args):
                 row["db_name"],
                 row["instructions"],
                 row["k_shot_prompt"],
+                row["glossary"],
+                row["table_metadata_string"],
+                row["prev_invalid_sql"],
+                row["prev_error_msg"],
                 public_data,
             ),
             axis=1,
