+++ eval/vllm_runner.py
@@ -4,7 +4,7 @@ import sqlparse
 from vllm import LLM, SamplingParams
 from eval.eval import compare_query_results
 import pandas as pd
-from utils.pruning import prune_metadata_str
+from utils.gen_prompt import generate_prompt
 from utils.questions import prepare_questions_df
 from utils.creds import db_creds_all
 import time
@@ -12,37 +12,6 @@ import torch
 from transformers import AutoTokenizer
 from tqdm import tqdm
 from utils.reporting import upload_results
-
-
-def generate_prompt(
-    prompt_file,
-    question,
-    db_name,
-    instructions="",
-    k_shot_prompt="",
-    glossary="",
-    table_metadata_string="",
-    public_data=True,
-):
-    with open(prompt_file, "r") as f:
-        prompt = f.read()
-    question_instructions = question + " " + instructions
-
-    if table_metadata_string == "":
-        pruned_metadata_str = prune_metadata_str(
-            question_instructions, db_name, public_data
-        )
-    else:
-        pruned_metadata_str = table_metadata_string
-
-    prompt = prompt.format(
-        user_question=question,
-        instructions=instructions,
-        table_metadata_string=pruned_metadata_str,
-        k_shot_prompt=k_shot_prompt,
-        glossary=glossary,
-    )
-    return prompt
 
 
 def run_vllm_eval(args):
@@ -95,6 +64,8 @@ def run_vllm_eval(args):
                 "k_shot_prompt",
                 "glossary",
                 "table_metadata_string",
+                "prev_invalid_sql",
+                "prev_error_msg",
             ]
         ].apply(
             lambda row: generate_prompt(
@@ -105,6 +76,8 @@ def run_vllm_eval(args):
                 row["k_shot_prompt"],
                 row["glossary"],
                 row["table_metadata_string"],
+                row["prev_invalid_sql"],
+                row["prev_error_msg"],
                 public_data,
             ),
             axis=1,
